+++ ServerBrowser/ServerRow.cs
@@ -26,22 +26,15 @@ namespace ServerBrowser
     private int isModified;
     internal bool QueryPlayers { get; set; }
     internal bool QueryRules { get; set; }
-    internal GameExtension GameExtension { get; private set; }
+    internal GameExtension GameExtension { get; set; }
     public GeoInfo GeoInfo { get; set; }
     public string CachedName { get; set; }
 
-    public ServerRow(IPEndPoint ep, GameExtension extension)
+    public ServerRow(IPEndPoint ep, GameExtension extension, ServerInfo info = null)
     {
       this.EndPoint = ep;
       this.GameExtension = extension;
-      this.PlayerCount = new PlayerCountInfo(this);
-    }
-
-    public ServerRow(ServerInfo info, GameExtension extension)
-    {
       this.ServerInfo = info;
-      this.EndPoint = info.EndPoint;
-      this.GameExtension = extension;
       this.PlayerCount = new PlayerCountInfo(this);
     }